requirements.txt
ccxt==4.3.35
telethon==1.28.5
aiohttp==3.9.1
rich==13.7.0
python-dotenv==1.0.0
pandas==2.0.3
numpy==1.24.3
